Refactor: Separate search highlight and implement in popup page
