+++ frontend/src/pages/IndexPage.tsx
@@ -4,6 +4,7 @@ import "../index.css";
 import QueryInput from "../QueryInput";
 import { GifPreviewSingle } from "../GifPreviewSingle";
 import { config } from "..";
+import { SearchHighlight } from "../components/SearchHighlight";
 
 const IndexPage: Component = () => {
     const [gifs, setGifs] = createSignal([] as Gif[]);
@@ -12,6 +13,9 @@ const IndexPage: Component = () => {
     return (
         <>
             <div class="content-header">
+                <h2>Home page</h2>
+            </div>
+            <div class="content-content">
 
                 <QueryInput setGifs={setGifs} setQuery={setQuery} />
 
@@ -27,45 +31,7 @@ const IndexPage: Component = () => {
                                 <a href={`/gifs/${gif.id}`}>
                                     <div style="display: grid; grid-template-rows: auto auto">
                                         <GifPreviewSingle gif={gif} tryForceCache />
-                                        <For each={gif.tags}>
-                                            {(tag, i) => {
-                                                //SEARCH HIGHLIGHTING
-                                                //First see if the search is present in the tag
-                                                const search = tag.match(query())
-                                                if (!search) { //If not then dont return it
-                                                    return (<></>
-                                                    )
-                                                }
-                                                //If so then split it into parts. (Could be multiple parts matching the query!)
-                                                const parts = tag.split(search[0])
-                                                    .reduce((list, elem, i) => {
-                                                        list.push(elem);
-                                                        if (i == list.length - 1)
-                                                            list.push(search[0]);
-                                                        return list;
-                                                    }, []);
-                                                console.log(parts)
-                                                return (
-                                                    <span class="tag">
-                                                        <For each={parts}>
-                                                            {(part, i) => (
-                                                                <span>
-                                                                    <Show when={part == search}>
-                                                                        <span style="background-color: rgba(255, 255, 0, 0.5)">
-                                                                            {part}
-                                                                        </span>
-                                                                    </Show>
-                                                                    <Show when={part != search}>
-                                                                        {part}
-                                                                    </Show>
-                                                                </span>
-                                                            )}
-                                                        </For>
-                                                    </span>
-                                                )
-                                            }
-                                            }
-                                        </For>
+                                        <SearchHighlight gif={gif} query={query()}></SearchHighlight>
                                     </div>
                                 </a>,
                             );
